Making the commands invoke serially
The original version could spawn far far too many processes.
This revision limits it to a single process at a time.

I'll probably do a parallel version of this by having a
service which partitions the JSON array, and then hands
off a subset of the array to different partitions.